Split assets and gameplay helpers into ui package
